add deltalake_storage_options to filesystem config
